re-format ui

+++ microtrafficsim-ui/src/main/java/microtrafficsim/ui/utils/Utils.java
@@ -14,7 +14,7 @@ import java.io.PrintStream;
 
 
 public class Utils {
-  public static void asyncScreenshot(RenderContext context) {
+    public static void asyncScreenshot(RenderContext context) {
         new Thread(() -> {
             JFileChooser chooser = new JFileChooser();
             chooser.setFileFilter(new FileFilter() {
@@ -31,48 +31,47 @@ public class Utils {
                     String extension = null;
 
                     String s = f.getName();
-                    int i = s.lastIndexOf('.');
+                    int    i = s.lastIndexOf('.');
 
-                    if (i > 0 &&  i < s.length() - 1)
-                        extension = s.substring(i+1).toLowerCase();
+                    if (i > 0 && i < s.length() - 1)
+                        extension = s.substring(i + 1).toLowerCase();
 
-                    if (extension == null) return false;
+                    if (extension == null)
+                        return false;
 
                     switch (extension) {
-                    case "png":     return true;
-                    default:        return false;
+                    case "png": return true;
+                    default:    return false;
                     }
                 }
             });
 
             int action = chooser.showSaveDialog(null);
-        
+
             if (action == JFileChooser.APPROVE_OPTION) {
                 File file = chooser.getSelectedFile();
                 if (file.exists()) file.delete();
 
                 try {
                     file.createNewFile();
-                } catch (IOException e) {
-                    return;
-                }
+                } catch (IOException e) { return; }
 
                 context.addTask(c -> {
                     try {
                         FramebufferUtils.writeFramebuffer(c.getDrawable(), "png", file);
                     } catch (IOException e) {
                         /* ignore if we can't write to the file and clean up */
-                        if (file.exists())
-                            file.delete();
+                        if (file.exists()) file.delete();
                     }
                     return null;
                 });
             }
-        }).start();;
+        }).start();
+        ;
     }
-    
-  public static class DebugExceptionHandler implements UncaughtExceptionHandler {
-    
+
+    public static class DebugExceptionHandler implements UncaughtExceptionHandler {
+
         @Override
         public void uncaughtException(RenderContext context, Throwable exception) {
             if (exception instanceof ShaderCompileError)
@@ -81,11 +80,11 @@ public class Utils {
                 exceptionPrintf(System.err, (ShaderLinkError) exception);
             else
                 exception.printStackTrace();
-        
+
             // XXX: clean exit strategy?
             Runtime.getRuntime().halt(1);
         }
-        
+
         private void exceptionPrintf(PrintStream out, ShaderCompileError error) {
             out.println(error.toString());
             out.println("-- LOG -------------------------------------------------------------------------");
@@ -95,7 +94,7 @@ public class Utils {
             out.println("-- STACK TRACE -----------------------------------------------------------------");
             error.printStackTrace(out);
         }
-        
+
         private void exceptionPrintf(PrintStream out, ShaderLinkError error) {
             out.println(error.toString());
             out.println("-- LOG -------------------------------------------------------------------------");
